+++ packages/zapier-platform-cli/src/app-templates.js
@@ -13,4 +13,5 @@ module.exports = [
   'session-auth',
   'babel',
   'rest-hooks',
+  'files',
 ];